requirements.txt
django==5.0.2
python-dotenv==1.0.1
mysqlclient==2.2.4
djangorestframework==3.15.1
pytest==8.1.1
pytest-django==4.5.0
freezegun==1.1.0
attrs==21.4.0
cattrs>=22.2.0,<23
